- entity.* 패키지로 enum 통일 (type.* → entity.*)
- Report.java: UniqueConstraint + admin 처리 필드 병합
- ReportService.java: create/getMyReports/getMyReport(90) + getReports/getReport(91) 통합
- ReportRepository.java: 사용자 조회 쿼리(90) + 관리자 조회 쿼리(91) 통합
- CursorUtils 공통 모듈 적용 (inline cursor → CursorUtils)
- AbstractIntegrationTest: portone.webhook-secret 더미 값 추가

+++ src/main/java/com/chunbaetour/domain/shop/dto/response/ShopResponse.java
@@ -19,7 +19,7 @@ public record ShopResponse(
         String operatingHours,
         String closedDays,
         boolean isCertified,
-        double rating,
+        BigDecimal rating,
         int reviewCount,
         ShopStatus status
 ) {
@@ -38,7 +38,7 @@ public record ShopResponse(
                 shop.getOperatingHours(),
                 shop.getClosedDays(),
                 shop.isCertified(),
-                shop.getRating(),
+                BigDecimal.valueOf(shop.getRating()),
                 shop.getReviewCount(),
                 shop.getStatus()
         );
